[Breaking] Add placeholder failing tests
