+++ bedrock/base/tests/test_urls.py
@@ -1,0 +1,15 @@
+# This Source Code Form is subject to the terms of the Mozilla Public
+# License, v. 2.0. If a copy of the MPL was not distributed with this
+# file, You can obtain one at https://mozilla.org/MPL/2.0/.
+
+
+def test_healthz():
+    assert False, "Write me"
+
+
+def test_healthz_cron():
+    assert False, "Write me"
+
+
+def test_readiness():
+    assert False, "Write me"
